pyproject.toml
[project]
name = "sign-vq"
description = "Quantize Sign Language Poses"
version = "0.0.1"
authors = [
    { name = "Amit Moryossef", email = "amitmoryossef@gmail.com" }
]
readme = "README.md"
dependencies = [
    "pose-format>=0.3.2",
    "opencv-python",
    "tqdm",
    "vector-quantize-pytorch",
    "pytorch-lightning",
    "wandb",
    # to support wandb video logging
    "moviepy",
    "imageio",
    # used to normalize the data
    "mediapipe",
    # to support a huggingface dataset, not recommended
    "datasets",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pylint"
]

[tool.yapf]
based_on_style = "google"
column_limit = 120

[tool.pylint]
max-line-length = 120
disable = [
    "C0114", # Missing module docstring
    "C0115", # Missing class docstring
    "C0116", # Missing function or method docstring
    "C0415", # Import outside toplevel
]
good-names = ["i", "f", "x", "y"]

[tool.pylint.typecheck]
generated-members = ["torch.*", "numpy.*", "cv2.*"]

[tool.setuptools]
packages = [
    "sign_vq"
]

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["sign_vq"]
